requirements.txt
adafruit-ampy==1.0.7
attrs==22.1.0
bitstring==3.1.9
black==22.10.0
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
docopt==0.6.2
ecdsa==0.18.0
esp32-machine-emulator==1.1.3
esptool==4.4
exceptiongroup==1.0.4
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.2
platformdirs==2.5.4
pluggy==1.0.0
pycparser==2.21
pyparsing==3.0.9
pyserial==3.5
pytest==7.2.0
python-dotenv==0.21.0
reedsolo==1.5.4
six==1.16.0
tomli==2.0.1
typing_extensions==4.4.0
